+++ modules/documentation/docs/changelog.md
@@ -2,6 +2,11 @@
 
 ## Next Release
 
+- \[contracts\] Add `transactionHash` to withdrawal commitment
+- \[store\] Add `transferDefinition` to filter options on `getTransfers`
+- \[store\] Add `getWithdrawalCommitmentByTransactionHash` method
+
+
 ## 0.1.15-beta.12
 
 - \[contracts\] Deploy contracts to BSC testnet
@@ -9,6 +14,7 @@
 ## 0.1.15-beta.11
 
 - \[iframe\] Allow `chainAddresses` and `messagingUrl` to be passed into iframe's node creation
+
 ## 0.1.15-beta.10
 
 - \[router\] Improve router dashboards